Tackle Events portion of the homepage
