+++ versions/v6/static/scss/components/_ucf-events.scss
@@ -1,0 +1,68 @@
+.ucf-events-modern-date {
+  .ucf-event-row {
+    @include make-row();
+    flex-wrap: nowrap;
+    padding-bottom: $grid-gutter-width-base / 2;
+    padding-top: $grid-gutter-width-base / 2;
+
+    &:not(:last-child) {
+      border-bottom: $hr-border-color solid $border-width-thickest;
+    }
+  }
+
+  .ucf-event-when {
+    @include make-col-ready();
+    flex: 0 0 6rem;
+    max-width: 6rem;
+    padding-top: .25rem;
+    text-align: center;
+  }
+
+  .ucf-event-start-day {
+    font-size: 1.05em;
+    font-weight: 700;
+    letter-spacing: -.0125em;
+    line-height: 1;
+    text-transform: uppercase;
+  }
+
+  .ucf-event-start-date {
+    color: $brand-default;
+    font-family: $font-family-slab-serif;
+    font-size: 4em;
+    letter-spacing: -.0125em;
+    line-height: .8;
+  }
+
+  .ucf-event-start-month {
+    font-size: .85em;
+    font-weight: bold;
+    letter-spacing: -.0125em;
+    line-height: 1;
+    text-transform: uppercase;
+  }
+
+  .ucf-event-details {
+    @include make-col-ready();
+    flex: 1 1 auto;
+    max-width: 100%;
+    width: auto;
+  }
+
+  .ucf-event-title-wrapper {
+    margin-bottom: .5rem;
+    margin-top: .4rem;
+  }
+
+  .ucf-event-title {
+    color: $brand-secondary;
+    display: inline-block;
+    font-size: 1.1em;
+    font-weight: bold;
+    line-height: 1.4;
+  }
+
+  .ucf-event-description-wrapper {
+    font-size: .8em;
+  }
+}
